refactor(InputFile): 'use client'を持つ内部コンポーネントをclient/へ移動する
InputFileNative・InputFileMultiplyAppendable・FilePreviewDialogは
useState/useRef等を使うため'use client'が必須。これらをclient/直下に
集約し、client境界を持つモジュールとそうでないモジュール
（parts.tsx・style.ts・types.ts）を分離する。

いずれもindex.tsから公開されない内部コンポーネントで、hookも存在しない
ためclient/直下にフラット配置する（CLAUDE.mdの判断基準）。
