+++ packages/smarthr-ui/src/components/InputFile/client/index.ts
@@ -1,1 +1,3 @@
 export { DownloadAnchorButton } from './DownloadAnchorButton'
+export { InputFileNative } from './InputFileNative'
+export { InputFileMultiplyAppendable } from './InputFileMultiplyAppendable'
